+++ ProjectManagerPrototype_v1/projectmanager/SimpleTaskApp.java
@@ -1,4 +1,4 @@
-package src.projectmanagerprototype;
+package projectmanager;
 
 import java.util.Scanner;
