+++ ProjectManagerPrototype_v1/projectmanager/Task.java
@@ -1,4 +1,4 @@
-package src.projectmanagerprototype;
+package projectmanager;
 
 import java.io.Serializable;
